Refactor profile UI by extracting components and improving layout
* refactor: Extract `StatsCard` and related sub-components into a dedicated file for better modularity
* feat: Move `ProfileLoadingSkeleton` to a separate file and refine skeleton sub-views for the profile header and stats
* refactor: Relocate `UnauthenticatedContent` to the `components` package and update its layout to use a weighted column for improved footer positioning
* chore: Clean up `ProfileScreen` by removing local component definitions and updating imports to use the new component structure
