+++ features/profile/ui/src/main/kotlin/com/thomaskioko/tvmaniac/profile/ui/components/UnauthenticatedContent.kt
@@ -1,10 +1,11 @@
-package com.thomaskioko.tvmaniac.profile.ui
+package com.thomaskioko.tvmaniac.profile.ui.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -23,6 +24,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.PreviewWrapper
@@ -31,6 +33,7 @@ import com.thomaskioko.tvmaniac.compose.components.FilledTextButton
 import com.thomaskioko.tvmaniac.compose.components.ThemePreviews
 import com.thomaskioko.tvmaniac.compose.components.TvManiacPreviewWrapperProvider
 import com.thomaskioko.tvmaniac.profile.presenter.model.ProfileLabels
+import com.thomaskioko.tvmaniac.profile.ui.sampleProfileLabels
 import com.thomaskioko.tvmaniac.testtags.profile.ProfileTestTags
 
 @Composable
@@ -42,63 +45,63 @@ internal fun UnauthenticatedContent(
 ) {
     Column(
         modifier = modifier
+            .fillMaxSize()
             .padding(contentPadding.calculateTopPadding())
-            .verticalScroll(rememberScrollState())
-            .padding(horizontal = 24.dp, vertical = 54.dp)
-            .statusBarsPadding(),
-        verticalArrangement = Arrangement.spacedBy(32.dp),
+            .statusBarsPadding()
+            .padding(horizontal = 24.dp, vertical = 54.dp),
     ) {
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Text(
-            text = labels.unauthenticatedTitle,
-            style = MaterialTheme.typography.displaySmall,
-            fontWeight = FontWeight.Bold,
-            lineHeight = MaterialTheme.typography.displaySmall.fontSize.times(1.2f),
-        )
-
         Column(
-            verticalArrangement = Arrangement.spacedBy(24.dp),
-            modifier = Modifier.fillMaxWidth(),
+            modifier = Modifier
+                .weight(1f)
+                .fillMaxWidth()
+                .verticalScroll(rememberScrollState()),
+            verticalArrangement = Arrangement.spacedBy(32.dp),
         ) {
-            FeatureItem(
-                icon = Icons.Outlined.Search,
-                title = labels.featureDiscoverTitle,
-                description = labels.featureDiscoverDescription,
+            Spacer(modifier = Modifier.height(8.dp))
+
+            Text(
+                text = labels.unauthenticatedTitle,
+                style = MaterialTheme.typography.displaySmall,
+                fontWeight = FontWeight.Bold,
+                lineHeight = MaterialTheme.typography.displaySmall.fontSize.times(1.2f),
             )
 
-            FeatureItem(
-                icon = Icons.Outlined.Tv,
-                title = labels.featureTrackTitle,
-                description = labels.featureTrackDescription,
-            )
+            Column(
+                verticalArrangement = Arrangement.spacedBy(24.dp),
+                modifier = Modifier.fillMaxWidth(),
+            ) {
+                FeatureItem(
+                    icon = Icons.Outlined.Search,
+                    title = labels.featureDiscoverTitle,
+                    description = labels.featureDiscoverDescription,
+                )
 
-            FeatureItem(
-                icon = Icons.AutoMirrored.Outlined.LibraryBooks,
-                title = labels.featureManageTitle,
-                description = labels.featureManageDescription,
-            )
+                FeatureItem(
+                    icon = Icons.Outlined.Tv,
+                    title = labels.featureTrackTitle,
+                    description = labels.featureTrackDescription,
+                )
 
-            FeatureItem(
-                icon = Icons.Outlined.AutoAwesome,
-                title = labels.featureMoreTitle,
-                description = labels.featureMoreDescription,
-            )
+                FeatureItem(
+                    icon = Icons.AutoMirrored.Outlined.LibraryBooks,
+                    title = labels.featureManageTitle,
+                    description = labels.featureManageDescription,
+                )
+
+                FeatureItem(
+                    icon = Icons.Outlined.AutoAwesome,
+                    title = labels.featureMoreTitle,
+                    description = labels.featureMoreDescription,
+                )
+            }
         }
-
-        Spacer(modifier = Modifier.height(2.dp))
 
         Column(
             verticalArrangement = Arrangement.spacedBy(20.dp),
-            modifier = Modifier.fillMaxWidth(),
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 32.dp),
         ) {
-            Text(
-                text = labels.footerDescription,
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                lineHeight = MaterialTheme.typography.bodySmall.fontSize.times(1.5f),
-            )
-
             FilledTextButton(
                 modifier = Modifier
                     .fillMaxWidth()
@@ -113,15 +116,20 @@ internal fun UnauthenticatedContent(
                     Text(text = labels.signInButton)
                 },
             )
-        }
 
-        Spacer(modifier = Modifier.height(32.dp))
+            Text(
+                text = labels.footerDescription,
+                style = MaterialTheme.typography.bodySmall,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                lineHeight = MaterialTheme.typography.bodySmall.fontSize.times(1.5f),
+            )
+        }
     }
 }
 
 @Composable
 private fun FeatureItem(
-    icon: androidx.compose.ui.graphics.vector.ImageVector,
+    icon: ImageVector,
     title: String,
     description: String,
     modifier: Modifier = Modifier,
